integrated histogram stuff into image.cpp
